# Optimized at_c and back.