--- a/brigand/sequences/back.hpp
+++ b/brigand/sequences/back.hpp
@@ -6,9 +6,10 @@
 =================================================================================================**/
 #pragma once
 
-#include <brigand/types/voidp.hpp>
+#include <brigand/sequences/at.hpp>
 #include <brigand/sequences/list.hpp>
 #include <brigand/sequences/last_element.hpp>
+#include <brigand/sequences/size.hpp>
 
 namespace brigand
 {
@@ -27,22 +28,8 @@ namespace brigand
   template<class L, class... T>
   using push_back = typename detail::push_back_impl<L, T...>::type;
 
-  // back
-  namespace detail
-  {
-    template<class L> struct back_impl;
-
-    template<template<class...> class L, class T, class... Ts>
-    struct back_impl<L<T, Ts...>>
-    {
-      template<class U> static U back(typename voidp<Ts>::type..., U*);
-      using type = decltype(back(nullptr, static_cast<Ts*>(nullptr)...));
-    };
-
-  }
-
   template <class L>
-  using back = typename detail::back_impl<L>::type;
+  using back = at_c<L, size<L>::value-1>;
 
   // pop back
   namespace detail