--- a/brigand/types/voidp.hpp
+++ b/brigand/types/voidp.hpp
@@ -1,0 +1,12 @@
+/*==================================================================================================
+  Copyright (c) 2015 Edouard Alligand and Joel Falcou
+
+  Distributed under the Boost Software License, Version 1.0.
+  (See accompanying file LICENSE.md or copy at http://boost.org/LICENSE_1_0.txt)
+=================================================================================================**/
+#pragma once
+
+namespace brigand
+{
+  template<typename T> struct voidp { using type = void const *; };
+}