… brigand/sequences/last_element.hpp