@porkybrain 2016-03-06
might be another #ifdef candidate. MSVC2013's alias support sucks!